Add from_expanded compiler option
This feature would be very useful for the Elixir compiler.

Recently, Elixir reimplemented almost all validations performed by
erl_lint and introduced some more. Reimplementing those directly in the
elixir compiler allows improving the quality of warnings since they
occur closer to the actual source. Unfortunately this means the work is
done twice.
From the point of view of the Elixir compiler, parse transforms and
record expansion passes also only add overhead.

The erlang compiler contains already an intermediate form that can be
emitted from the compiler after the erl_lint, parse transform and record
expansion passes called 'E'. We propose adding a `from_expanded` option
to the compiler that would allow accepting this form as an input.

In most modules we've benchmarked using the `from_expanded` option in
the Elixir compiler translated to 5-10% of speedups in macro benchmarks,
which is siginificant enough to warrant this proposal in our opinion.

The alternative would be to call the `v3_core` module directly and
continue compilation with the `from_core` option. We would prefer an
officially supported feature, though.