Fix dbg for remote nodes
dbg uses the new trace sessions and stores them in the main dbg server
on the node that's doing the tracing. However, trace sessions are
reference-counted NIF resources - this means that when the last reference
is GC'ed on the node that owns the resource, the resouce is destroyed.

In case of dbg, this meant that as soon as the first GC happened
on the relay process, the session would be destroyed and tracing would
silently stop. Furthermore, any subsequent requests to amend
trace patterns, etc would fail for the remote nodes with badarg
in trace module's functions.

This fixes the issue by making the relay process hold onto the session
reference for as long as the process is active. This solves the issue.

The amended test fails before applying this change and succeeds after.