Apply a post-link optimization with BOLT
Summary:
This adds a a step to our build that runs BOLT on different variants
of the BEAM.

When applying BOLT using profile data for a specific service, CPU
utilization improved by ~4% on smaller services and ~6% on large
services.

To get those gains today, we would need to either manually collect
profiles for each service, increasing the complexity of OTP releases,
or make improvements to the build of OTP to use an automated system
for collecting profile and applying BOLT as part of the fbpkg build.

This change takes a pragmatic approach of collecting profile data by
running the dialyzer.  While this generates only a ~3% improvement on
chatd, it does so without any operational overhead.  In the meantime
we can work on the improvements to the build to enable the automated
collection of profiles in anticipation of growing that small win into
something larger.

Test Plan: Run tests and canary

Reviewers: skatepalli, yingyingxiao, zerth, #whatsapp_clr

Reviewed By: skatepalli

Subscribers: jordalgo, slinger, samiur

Differential Revision: https://phabricator.intern.facebook.com/D61157606

Tasks: T107504422

Tags: commitClose

[otp] update bolt build script to us local copy of perf data for devvm builds

Summary:
This diff addresses the below issues:
- Added parsetools to fix errors in dialyzer command.
- As we cannot run perf command on devvm, added a way to bypass running perf command.
- For devvm we will use a local of perf.data.
- Once I fix sandcastle automation (D62836976, D62835122), then we should be able to generate diffs which update perf.data file for devvm work flow.

Test Plan: https://www.internalfb.com/sandcastle/workflow/2256303413316868349

Reviewers: cshapiro, zeyu, yingyingxiao, #whatsapp_clr

Reviewed By: zeyu, yingyingxiao

Subscribers: yingyingxiao, samiur

Differential Revision: https://phabricator.intern.facebook.com/D62765722

Tasks: T194266694

Privacy Context Container: L1141030

[otp] Support bolt build on devvm

Summary:
As part of enabling bolt for OTP we run `perf` command to capture perf data on dialyzer command during build process, but devvm cannot run perf command, so we download perf data from manifold. The perf data is uploaded to manifold as part of sandcastle command and saves the information in an executable, which is run as part of devvm build.

Sample diff generated by the sandcastle command. D63229690

Sandcastle command to generate diff. D62835122

Test Plan: Manual test.

Reviewers: cshapiro, zeyu, yingyingxiao, #whatsapp_clr

Reviewed By: zeyu, yingyingxiao

Subscribers: yingyingxiao

Differential Revision: https://phabricator.intern.facebook.com/D63228940

Privacy Context Container: L1141030

Address issues with running the bolt.sh script

Summary:
The bolt.sh script does not reliably detect cases when perf can be run
successfully, such as when it is invoked on a devbig, and it always
falls back to downloading profile data from Manifold.  To resolve this
issue, the following changes were made

  1) Running on sandcastle is deteced by checking to see if the
     environment variable $SANDCASTLE is defined.  This matches the
     behavior in libfb, see fbcode/libfb/py/environment.py for details.

  2) A devvm is detected by checking for a hostname that starts with
     of "devvm".  Not all devvms have a virtualized PMU making perf(1)
     unreliable on devvms in general.

  3) Quote newly added variables to avoid unwanted globbing after
     their expansion.

  4) Rename PERF_FILE_NAME to PERF_DATA to signal that we are talking
     about the perf.data file as opposed to the perf command itself or
     another file related to perf(1).

Finally, the warning message is clarified.

Test Plan: Succesfully ran a build on a devbig

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D67428029

fix bolt build script for sandcastle.

Summary: Shell script need to use '-z' instead of '-n' to detect if the build is running in sandcastle

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: